pyproject.toml
[project]
name = "notificatu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "beautifulsoup4==4.13.4",
    "bs4==0.0.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "comm==0.2.2",
    "debugpy==1.8.13",
    "decorator==5.2.1",
    "docker>=7.1.0",
    "executing==2.2.0",
    "fastexcel==0.13.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.0.2",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "matplotlib-inline==0.1.7",
    "nest-asyncio==1.6.0",
    "numpy==2.2.4",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pip==24.0",
    "platformdirs==4.3.7",
    "polars-lts-cpu>=1.33.1",
    "prompt-toolkit==3.0.50",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==19.0.1",
    "pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyzmq==26.4.0",
    "qrcode>=8.2",
    "qrcode-terminal>=0.8",
    "requests==2.32.3",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "stack-data==0.6.3",
    "tornado==6.4.2",
    "traitlets==5.14.3",
    "typing-extensions==4.14.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "xlsxwriter>=3.2.4",
]
